Move BoardView, PersonInfo and QuoteInfo to pages dir and append Page to filenames.
